backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.95.2
uvicorn==0.22.0

# Database ORM and driver
sqlalchemy==2.0.15
alembic==1.11.1
psycopg2-binary==2.9.6

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Data validation
pydantic==1.10.8
email-validator==2.0.0.post2

# Testing
pytest==7.3.1
pytest-cov==4.1.0
httpx==0.24.1

# API documentation
openapi-schema-pydantic==1.2.4

# OpenAI integration (for future phases)
openai==0.27.8

# LangChain for AI orchestration (for future phases)
langchain==0.0.200

# Utilities
python-dotenv==1.0.0
tenacity==8.2.2
